Assignment 2 überarbeitete Aufgabe
