+++ Assignment2/OrderSystem/src/logic/OrderService.java
@@ -1,8 +1,8 @@
 package logic;
 
 import model.Item;
-import model.Order;
-import model.SimpleItemFactory;
+//import model.Order;
+import factories.SimpleItemFactory;
 import ui.OrderViewer;
 
 import java.time.LocalDateTime;
@@ -11,7 +11,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 public class OrderService {
 
-	private Order order = new Order();
+	//private Order order = new Order();
 	private OrderViewer orderViewer = new OrderViewer();
 	private SimpleItemFactory simpleItemFactory = new SimpleItemFactory();
 
@@ -22,8 +22,8 @@ public class OrderService {
 			input = Input.readInt();
 			switch ( input ) {
 				case 0: break ;
-				case 1: simpleItemFactory.createProduct(order); break ;
-				case 2: simpleItemFactory.createService(order); break ;
+				case 1: simpleItemFactory.createProduct(simpleItemFactory); break ;
+				case 2: simpleItemFactory.createService(simpleItemFactory); break ;
 				default: System.out.println("invalid"); break ;
 			}
 		} while( input != 0 );
@@ -34,7 +34,7 @@ public class OrderService {
 	private void sortItems() {
 		Comparator<Item> byPrice =
 				Comparator.comparingInt(Item::getPrice);
-		order.getItems().sort(byPrice);
+		simpleItemFactory.getItems().sort(byPrice);
 	}
 
 	public void setSimpleItemFactory(SimpleItemFactory simpleItemFactory) {
@@ -43,22 +43,22 @@ public class OrderService {
 
 	private void finishOrder() {
 		AtomicInteger sum = new AtomicInteger();
-		order.getItems().forEach(item -> {
+		simpleItemFactory.getItems().forEach(item -> {
 			orderViewer.printItemPrice(item,formatPrice(item.getPrice()));
 			sum.addAndGet(item.getPrice());
 		});
 
-		order.setSum(sum.get());
-		order.setCheckoutTime(LocalDateTime.now());
+		simpleItemFactory.setSum(sum.get());
+		simpleItemFactory.setCheckoutTime(LocalDateTime.now());
 
-		orderViewer.printOrder(order, formatPrice(order.getSum()));
+		orderViewer.printOrder(simpleItemFactory, formatPrice(simpleItemFactory.getSum()));
 
 		emptyCart();
 		menuLoop();
 	}
 
 	private void emptyCart(){
-		order.getItems().clear();
+		simpleItemFactory.getItems().clear();
 	}
 
 	private String formatPrice(int priceInCent) {
