--- Assignment2/OrderSystem/src/model/ItemFactory.java
@@ -1,6 +1,0 @@
-package model;
-
-public interface ItemFactory {
-    void createProduct(Order order);
-    void createService(Order order);
-}
