+++ Assignment2/OrderSystem/src/ui/OrderViewer.java
@@ -1,5 +1,6 @@
 package ui;
 
+import factories.SimpleItemFactory;
 import model.Item;
 import model.Order;
 /**
@@ -15,29 +16,9 @@ public class OrderViewer {
         System.out.println("(2) Order service");
     }
 
-//    public void orderProduct(Order order) {
-//        System.out.println("Name: ");
-//        String productName = Input.readString();
-//        System.out.println("Unit price (in cents): ");
-//        int productPrice = Input.readInt();
-//        System.out.println("Quantity: ");
-//        int productQuantity = Input.readInt();
-//        order.getItems().add(new Product(productName, productPrice, productQuantity));
-//    }
-
-//    public void orderService(Order order) {
-//        System.out.println("Service type: ");
-//        String serviceName = Input.readString();
-//        System.out.println("Number of persons: ");
-//        int servicePersons = Input.readInt();
-//        System.out.println("Hours: ");
-//        int serviceHours = Input.readInt();
-//        order.getItems().add(new Service(serviceName, servicePersons, serviceHours));
-//    }
-
-    public void printOrder(Order order, String formattedPrice){
+    public void printOrder(SimpleItemFactory simpleItemFactory, String formattedPrice){
         System.out.println("Sum: "+ formattedPrice);
-        System.out.println("Checkout time: " + order.getCheckoutTime());
+        System.out.println("Checkout time: " + simpleItemFactory.getCheckoutTime());
         System.out.println("Session ended!\n");
     }
